refactor(bash): create module for bash factory
